requirements.txt
requests==2.21.0
ipinfo==3.0.0
Adafruit_Blinka==3.0.1
astral==1.10.1
adafruit_circuitpython_neopixel==3.4.0
pytz==2019.1
numpy==1.17.1
pyowm==2.10.0
board==1.0
flask==1.1.1
flask_cors
